Create script2.sh
The script will install Apache2 and then create a user. Afterward, it will configure the Apache2 virtual host with the path of the user so that the document root is set to the user's path. Finally, the script will pull the index page from GitHub and place it in the document root path.